Twice the bits, twice the blood? (64-bit build)
Once again, these patches allow the game to build on 64-bit, plus
a shell script to choose which binary to launch. makeit.sh is
included because changing the BINDIR didn't seem to have any effect
within the makefile itself, so I use it to put 32- and 64-bit
object files in different locations. It's a convenience thing,
simply using "make" will still work.

+++ RSPiX/Src/BLUE/unix/Bmouse.cpp
@@ -41,11 +41,11 @@ extern int sdlWindowHeight;
 
 typedef struct
 	{
-	short	sX;
-	short	sY;
-	short	sButton;
-	long	lTime;
-	short	sType;
+	int16_t	sX;
+	int16_t	sY;
+	int16_t	sButton;
+	int32_t	lTime;
+	int16_t	sType;
 	} RSP_MOUSE_EVENT, *PRSP_MOUSE_EVENT;
 
 #define MAX_EVENTS	256
@@ -62,7 +62,7 @@ extern bool mouse_grabbed;
 ///////////////////////////////////////////////////////////////////////////////
 // Module specific (static) globals.
 ///////////////////////////////////////////////////////////////////////////////
-static short				ms_sCursorShowLevel	= 0;
+static int16_t				ms_sCursorShowLevel	= 0;
 
 ///////////////////////////////////////////////////////////////////////////////
 // Functions.
@@ -79,9 +79,9 @@ static int MouseWheelState = 0;
 // 
 ///////////////////////////////////////////////////////////////////////////////
 extern void rspGetMouse(
-		short* psX,				// Current x position is returned here (unless NULL)
-		short* psY,				// Current y position is returned here (unless NULL)
-		short* psButton)		// Current button status is returned here (unless NULL)
+		int16_t* psX,				// Current x position is returned here (unless NULL)
+		int16_t* psY,				// Current y position is returned here (unless NULL)
+		int16_t* psButton)		// Current button status is returned here (unless NULL)
 	{
 
     if (!mouse_grabbed)
@@ -115,7 +115,7 @@ extern void rspGetMouse(
 
 extern void Mouse_Event(SDL_Event *event)
 {
-	static short	sEventIndex	= 0;
+	static int16_t	sEventIndex	= 0;
 
 
     if (!mouse_grabbed)
@@ -127,7 +127,7 @@ extern void Mouse_Event(SDL_Event *event)
     pme->lTime = SDL_GetTicks();
     pme->sType = event->type;
 
-    static short buttonState = 0;
+    static int16_t buttonState = 0;
 
 	bool bQueueMouseWheelRelease = false;
 
@@ -215,8 +215,8 @@ extern void Mouse_Event(SDL_Event *event)
 // 
 ///////////////////////////////////////////////////////////////////////////////
 extern void rspSetMouse(
-		short sX,				// New x position.
-		short sY)				// New y position.
+		int16_t sX,				// New x position.
+		int16_t sY)				// New y position.
 	{
         if (!mouse_grabbed)
             return;  // drop mouse events if input isn't grabbed.
@@ -229,17 +229,17 @@ extern void rspSetMouse(
 // This function tosses out any events ahead of the last event in the queue!
 // 
 ///////////////////////////////////////////////////////////////////////////////
-extern short rspGetLastMouseEvent(	// Returns 0 if no event was available, non-zero otherwise
-	short*	psX,						// Event's X position is returned here (unless NULL)
-	short*	psY,						// Event's Y position is returned here (unless NULL)
-	short*	psButton,				// Event's button status is returned here (unless NULL)
-	long*		plTime,					// Event's time stamp returned here (unless NULL)
-	short*	psType /*= NULL*/)	// Event's type (as per OS) is returned here (unless NULL)
+extern int16_t rspGetLastMouseEvent(	// Returns 0 if no event was available, non-zero otherwise
+	int16_t*	psX,						// Event's X position is returned here (unless NULL)
+	int16_t*	psY,						// Event's Y position is returned here (unless NULL)
+	int16_t*	psButton,				// Event's button status is returned here (unless NULL)
+	int32_t*		plTime,					// Event's time stamp returned here (unless NULL)
+	int16_t*	psType /*= NULL*/)	// Event's type (as per OS) is returned here (unless NULL)
 	{
-	short	sRes	= TRUE;	// Assume success.
+	int16_t	sRes	= TRUE;	// Assume success.
 
 	PRSP_MOUSE_EVENT	peEvent;
-	short					sNumEvents	= ms_qmeEvents.NumItems();
+	int16_t					sNumEvents	= ms_qmeEvents.NumItems();
 
 	// Are there any events?
 	if (sNumEvents > 0)
@@ -277,14 +277,14 @@ extern short rspGetLastMouseEvent(	// Returns 0 if no event was available, non-z
 // Returns 0 on success.
 // 
 ///////////////////////////////////////////////////////////////////////////////
-extern short rspGetMouseEvent(	// Returns 0 if no event was available, non-zero otherwise
-	short*	psX,						// Event's X position is returned here (unless NULL)
-	short*	psY,						// Event's Y position is returned here (unless NULL)
-	short*	psButton,				// Event's button status is returned here (unless NULL)
-	long*		plTime,					// Event's time stamp returned here (unless NULL)
-	short*	psType /*= NULL*/)	// Event's type (as per OS) is returned here (unless NULL)
+extern int16_t rspGetMouseEvent(	// Returns 0 if no event was available, non-zero otherwise
+	int16_t*	psX,						// Event's X position is returned here (unless NULL)
+	int16_t*	psY,						// Event's Y position is returned here (unless NULL)
+	int16_t*	psButton,				// Event's button status is returned here (unless NULL)
+	int32_t*		plTime,					// Event's time stamp returned here (unless NULL)
+	int16_t*	psType /*= NULL*/)	// Event's type (as per OS) is returned here (unless NULL)
 	{
-	short	sRes	= TRUE;	// Assume success.
+	int16_t	sRes	= TRUE;	// Assume success.
 
 	PRSP_MOUSE_EVENT	peEvent	= ms_qmeEvents.DeQ();
 	if (peEvent != NULL)
@@ -376,7 +376,7 @@ extern void rspUnshieldMouseCursor(void)
 // Reports current mouse cursor show level.
 // 
 ///////////////////////////////////////////////////////////////////////////////
-short rspGetMouseCursorShowLevel(void)	// Returns current mouse cursor show level:
+int16_t rspGetMouseCursorShowLevel(void)	// Returns current mouse cursor show level:
 													// Positive indicates cursor is shown.
 													// Non-positive indicates cursor is hidden.
 	{
@@ -389,7 +389,7 @@ short rspGetMouseCursorShowLevel(void)	// Returns current mouse cursor show leve
 // 
 ///////////////////////////////////////////////////////////////////////////////
 void rspSetMouseCursorShowLevel(	// Returns nothing.
-	short sNewShowLevel)				// In:  Current mouse cursor show level:        
+	int16_t sNewShowLevel)				// In:  Current mouse cursor show level:        
 											// Positive indicates cursor is shown.     
 											// Non-positive indicates cursor is hidden.
 	{
